The app started up. The data base was created and empty cursors were queried.
I had a problem with the joins and decided to make all of the field names unique so I didn't have to set aliases in the joins. This caused problems with the CursorAdapter because it wants an "_id" field. I added code to the Cursors that are used by adapters to alias the id column in the cursor to be "_id".