Data base and list view are working
Now there aren't book views. I changed the views to a tag list and will supply filters to do the views. It simplified quite a lot of stuff.

I added a source field for the source of the book data. The volumeId now only needs to be unique for a source.