Get Delete working
Delete is an ActionMode that is started by the Books fragment.

I needed to move the options menu to from the activity to the BookFragment

I needed to get selection working. I added a member to BookAndAuthors to keep the selected state. I add methods to the BookRepository to select and deselect individual books and to select or deselect all books. There is a LiveData<> property that holds selection changes which gets updated by the selection methods. I also keep track of the number of books selected.

I noticed some bugs in the BookDatabase and fixed them, too. When I changed a new book, I didn't delete the old authors, categories and tags. I probably should clear the thumbnails for the book, too.

I changed the book adapter to only update the displayed child in onBindViewHolder. The fragment listens to the selection changes and makes that work.