Change thumbnail handling
I moved the thumbnail handler into the Book DAO of the database. There I can use the Transaction mechanism to synchornize thumbnail fetches and book updates. I was having matching requests from the RecyclerView adapter to the proper database when creating the cache file.

This is how it works:

If the thumbnail file exists, then we load the bitmap from the file and return what we get.

If the file doesn' exist, we query the database for the thumbnail url and create an empty file in a transaction. Outside of a transaction we download the file into a tmp file. Finally inside of a transaction we check to see if the cache file we created still exists and move the temp file to the cache file.

When we delete or update a book we delete the cache file so it is downloaded again.

I also updated to Kotlin 1.4.10 to get funcional interfaces working.