Switch to Paging for the book database
Hopefully this will let me display search results from online more easily.

This was a big change because I couldn't find a way to get the content directly from the paging streams. I moved the selection code from the book repository to the view model and put it in a base class so we can support other view models. I connected the view model to the activity, so it can be shared among other fragments.

I also had to change code in the data base for deleting things, because I use only the ids for the deletion and because the selection set can be a set of selected ids, or unselected ids.

I changed the delete modal action to delete immediately if something is selected. Otherwise it starts the delete model action.