Tags are working
Tags are added/removed/edited in a bottom sheet managed by TagsFragment.

The book list has actions to tag, remove tags and replace tags for selected books. The currently selected tags in the bottom sheet are added, removed or replaced.

The book scanner will add the currently selected tags to books it scans.

The TagDao was getting complicated because it operated on two tables, so I broke it into two daos one for each table.