Move Options Menu UI to right side drawer
The books fragment finds all clickable buttons in the drawer and sets a common click listener that pushes clicks through the options menu code.

The buttons in the drawer are enabled and disabled to prevent the modal tag and modal delete actions from happening, but I left them in case I need them later.

I added some icons for the drawer and removed other icons that aren't used any longer.