Add Comments and refactor
Refactor BookFilter:

Move database columns to Column enum constructor along with the resource id for the enum display name. The default implementations of addSelect() and addOrder() use the columns from the constructor.

Change isSame to shouldAddSeparator.

Add getSeparatorValue() to get the value displayed in separators. It defaults to getValue() but can be different.

Removed unused methods from BookRepository