Comments and refactor
Made tag selection color public so it can be set by the tag fragment

Removed the description visibility change. Just set the text to ""

Move the TagsAdapter constructor inside of the TagViewModel constructor

Change the tag edit/add conflict dialog a little. Trim the name and description and keep the edit dialog open if the user replies No.