Refactor for BookFilter Serialization
I refactored BookFilter some to help with my serialization scheme. I am concerned about versioning the BookFilter and so I have decided to use different classes for each version. The class BookFilter will always refer to the current version, and older version will be named with a trailing version number like BookFilter0, BookFilter1. I will get the deserializer for the specific version and use that to deserialize the data. Then code will convert the old version classes to the current version.

With that in mind, I made lots of changes to reduce the footprint of the objects we serialize:

Move the enums from BookFilter to the file scope.

Put the companion object methods for BookFilter in its own class which the BookFilter companion object inherits.

Remove the methods from the Column enum and put them in a descriptor class that the enum references.