Filter UI
The filter UI is very similar to the order UI. It puts the UI in a table. The filter UI is more complex because we need to track values. The values are in a separate row of the table, but not in a TableRow. UIRow is an inner class of FfilterTable that handles some of the UI issues. Another difference is that the predicate spinner can change for each column in the column spinner.

I ran into an initialization order issue with the predicate data descriptor. The predicate enum values were being constructed before the data descriptors. I moved some code around to get everything initialized in the proper order.