Saved filters
The set of filters in the database are loaded and added to the navigation drawer. When they change the draw is updated.

The actions for the filter item is to navigate to the book fragment with the filter name as the argument. We check and prevent navigation to the book fragment using the current filter. The title is set from the filter if there is one.