Fix Scan recycler view selection
I discovered that I was losing selections in the Scan select books dialogs because I wasn't tracking the selected books outside of the populated views. I decided to change things around pretty drastically.

The ScanFragment now uses the BooksAdapter to display the books from books.google.com. I extracted the always visible portion of the books_adapter_book_item layout into another file which is included in books_adapter_book_item. I changed the BooksAdapter to check views for null so it can work with either the whole item or just the visible part.

I removed the Selection interface and use a lambda to handle the selection state. I changed where the thumnail cache is kept, so I can have multiple caches. Now thumbnails are in the thumbnails directory of the cache and thumbnails for the database are in the db subdirectory of thumbnails. The ScanFragment puts its thumbnails in a gb subdirectory which is deleted when the select books dialog is finished.