Add chips to tags in the book list
The chips can be used to add and remove tags from a single book.

I refactored the chip box and put all of the communication from the box to the client in a single Delegate interface. I supplied default implementations for the methods. I also handle closing and editing the chips with callbacks to control it. Most of the functions in the interface are suspend, so you can use coroutines to access the database, etc.

I also added a close icon when editing a chip to allow the edit to be canceled.

I decide to remove the context from the BooksAdapter and get it from the views.