Performance improvements
I moved code in BooksAdapter that deals with the view holder into the BookViewHolder class. I added code to cancel the thumbnail jobs when the view holder is recycled. This gets rid of jobs as you scroll. I also found that adding the Chips to all of the book details was slowing things down. So, I changed to code to only populate the details when they are visible. This got the scroll to work quickly and opening the details was still performant.

I change the Thumbnails.getThumbnail routine to return null if any exception happens, except for the CancellationException, when a thumbnail job is canceled.

Note that since the large thumbnail is in the detaill, we don't request it until the details are opened.