Fixed startup of book list
A changed notificationw as sent when adding the observer to build the flow for the book list view. This happend before the view was set and so the first flow was unfiltered, and then the proper filter was applied.

I changed the filter view LiveData to explicitly allow and initial null value. buildFlow doesn't do anything when the value is null, so the extra notification is ignored.