Include book filter in tag and delete modal actions
Before this change the book filter wasn't included, which could cause books that weren't visible in the book list to be changed.

I added a new abstract class that extends BuildQuery called SQLiteQueryBuilder. The new class adds methods to build from an iterator of OrderFields and FilterFields, and methods to get the SQLite query command and the SQLiteQuery. I added a methods to BuildQuery and BookFilter to create an SQLiteQueryBuilder.

When the BooksViewModel builds the flow for a filter. It builds the command and arguments needed to query the book ids for the filter and stores them in idFilter using a private setter.

The delete and tag actions pass the compiled filter into the database methods used to execute the actions. Finally the method that builds the queries to the database include the filter command in the queries.

I also added methods to count the number of tags and books that would be updated, so the counts in the verify dialogs would be correct. Because of that I changed those dialogs to coroutineAlertsand removed the dialog messages used when the count wasn't known. I still haven't updated the code that manages the menus.