Move selected and open into the book database
I added a flags column to the books and tags table to hold the selection state and open state. Each state is a single bit in the flags value.

I added two objects to the BookRepository to handle the flags. One is for books and one is for tags. I added a SelectionInterface to handle the selection in BaseViewModel. I also added an implementation that uses these objects to manage the selection. Because access to the database is suspended, I included unsuspended and suspended methods to update the selection. The unsuspended methods are asynchronous and launch a job. The suspending methods are synchronous. The selection object also holds LiveData for the current selected and item counts, which are used throughout.

For each search, the ScanFragment now adds deselects all books and adds the result selected. This allows the user to easily update the tags.

I added methods to generate queries using the flags value and to update the flags value. The bits can be cleared set or toggled.

I added a stats bar in the app bar to show the number of books and number of selected books.