ScanFragment sets filters to null
I realized that the book filters would still apply while the scan fragment was active, which would intefere with setting tags after books are added. So I have the scan fragment set the filter to null.

This meant that I needed to make the filter MutableLiveData public, and move the call to buildFlow() into the books fragment, so make sure changing the filter didn't trigger anything bad while the books fragment wasn't active.