Add OCR for ISBN Codes
It doesn't work very well and I am going to immediately remove it. I am adding it just to save it if I want it later.

I changed the code to use ImageCapture rather the ImageAnalyzer.

I added some code to interface with the Android GPU compute system. I use it to convert RGB to grayscale for the bar code scanner and OCR.

I added an empty scan view model, just so I could scope the scanner jobs with the view model scope.