Revamp the filter SQL builder to include isn't and hasn't
Remove selecting columns when using them in WHERE or ORDER clauses. I found out it wasn't necessary and it simplifies some other things.

Removed JOINs when filtering on authors, tags and categories. These now use a subquery with EXISTS to select the books that match.

Added isn't and hasn't predicates which negate the is and has predicates. These calculate the normal expression and either wraps it in a NOT ( ... ) or NOT ... depending on whether a sub-query was used or not.
Added sub-query builders for authors, tags, and categories and re