Add First BookFilter tests.
Test Single filter with empty values array

Test Single filter with randomly populated values array.

This is how the expected result is calculated:

We start the the sequence of books in the database and filter it. The filter takes the sequence of values from the book and the values from the array and performs the comparison based on the operator. The result either accepts or rejects the book in the sequence.