Add Single Order test and use debug databse for filter tests
We order by all columns both ascending and descending. We build the filter and then get the ordered books. First we check that the order of the books matches the order requested. Then we resort the books and include the book id as the final sort key. We also sort the expected books and verify that the lists are the same.

To sort the books and check the order we build a sequence of order lambdas that we interogate. We use the + operator for sequences add in the book id as the final sort key.

We also need to expand the expected book sequence to allow for the author, category and tag joins. We do this by mapping the books to a sequence of book sequences. For the non join columns the book sequence is just the book itself. For the join colums the booksequence is created by mapping the join list to a sequence of books. I needed to special case empty lists, which still include a single book. The sequence of book sequences is then flattened to a sequence of books.