Add multiple order and filter test along with a test for debugging
The multiple order and filter test fails and is ignored. Because of the failure, refactored the tests to see if I could make debugging them easier. The refactoring was done so that I could serialize the failing test to a string and display it with the error. Then I could use the string to construct the failing test and debug only that filter.

From the top-down I created a TestFilter class that holds the filter description and contains the method to test the filter. It overrides toString() to dump the serialized string for the filter description if it fails. TestFilter is a data class that contains a list of ordering a list of filtering. I added some secondary constructors to construct the filter from various sequences and lambdas.

I changed the filtering list to be a Triple that holds that column, predicate and values. I added a constructor to convert the column, predicate pair with a lambda to create the values to the Triple.

I made the ColumnValues, PredicateTests and PredicateValues classes enums so the serializer could handle them.