Add undo/redo to the Book Database
Undo adds two additional tables to the database to keep track of undo. It also adds HIDDEN flags to Book, Author, Category, Tag and View entities. Rows that have the hidden flag set are not part of the current database, but can be resored using undo or redo. Each undo transaction gets an transaction row in the undo table. Rows are marked as undo or redo, and are assigned an undo id when they are added. All of the individual operations for the transaction go in the operation table. The operations rows have an undo id and operation id. The undo id matches an operation to a transaction, and the operation id orders the operations within a transaction. There can be multiple rows with the same undo and operation id. In these cases the order shouldn't matter. Each operation gets a type, which is an ordinal of an enum in UndoRedoDao. The enum does the work of recording, undoing, redoing and discarding undo and redo operations.

Lots of method names in the database access objects were changed to ad "WithUndo", which indicates that undo records will be created if the method is called while recording. These methods call a record* method on one fo the OperatonType enum to record the operation to the database. Query methods were changed to filter out rows with the HIDDEN flag set.

The undo/redo code is in the UndoRedoDao in UndoRedo.kt.

The record methods check whether undo is recording, whether the operation succeeds to determine what to do. If undo is not recording the operation type, executes the real operation and returns. If undo is recording, undo executues the undoable operataion and if that succeeds records operation rows to undo the operation. Undoable operations in the book, authors, tags, categories and views tables general use the HIDDEN flag. In the link tables, the operation is performed, and the link is saved in the operation row.

Add and Update are special. The add is done and if it succeeds, the undo record is created. For updated, a copy is made of the row being updated, and it is deleted if the update fails.

When operations are discarded they delete any extra data they need from the database.

The method withUndo() {...}, runs the lamda argument while recording undo records. withUndo can be nested. The BookRepository usually calls withUndo() for each operation. The application can call it at a higher level to group multiple operations together. withUndo discards undo and redo transactions as needed and will reset undo ids occasionally.

While testing I discovered that the PageSource<>.forEach() method didn't work exactly right with the database. So I added some methods to get lists. I change the import export code to use those methods.

Some additional changes:

Update gradle to most recent version. Change the right margin. Add ifnull to the project dictionary.