Fix undo of view changes
There were two problems: The undo wasn't recorded correctly and the app wasn't changing the UI when the change was undone.

The recording error was a problem with the SQLite to record updates for views. The filter column wasn't being copied.

To fix the UI error, I added a method to get the current view with a LiveData and changed to code to respond to the LiveData changes.

I also discovered that I the registration and unregistration of the undo levels preference listener were wrong. I needed to make sure mDb was not null when the listener is registered.

I discovered that LiveData<> return values from the RoomDatabase do not need to be suspended. I changed this in a couple of places and filtered that out to where the return values are used.

I changed the test to catch the misrecording.

I ran into an unusual problem with an interface object declared this way Observer<List<UndoTransactionEntity>> { }. Kotlin was using a static instance of the object, which caused problems with registering the observer when switching between from one BooksFragement to a different one. Initializing it using object: worked right.

I learned about the switchMap LiveData Transformation and replaced my home grown CascadeLiveData class with that.