Change Database to Allow multiple ISBNs for a book
I added another pair of tables with ISBNs and links. The database now allows multiple books for an ISBN and multiple ISBNs for a book. I did it this way for a couple of reasons:

There was existing code that did that for other values in the database.
I decided ISBNs could be wrong, so allowing multiple books per ISBN keeps that from being an issue.