Start of printing
There is a print adapter and a layout system for layout out books on pages.

The overall flow is this:

When needed we layout the pages we are printing by laying out each book and then figuring out how to position it on the page and handling page breaks as need. A list of pages containing the location and id of each book on the page is built during this phase. The only tricky part of this process is making sure each field breaks along an asthetic boundary - like between lines. When a book is continued onto another column or page, the location is set to a negative value and clip rectangles are used to make sure that the correct content is shown.

When we print, we just loop through the pages lookup and layout each book, setup the clip rectangles and draw the fiels.

The layout is described by the LayoutDescription. The structure is similar to ConstraintView. You can have margins and constrain start, end, top and bottom to multiple other field's start, end, top and bottom. When multiple fields are used the constraint depends on what is being constrained

Vertical
Top constrains to the max value
Bottom constrains to the min value
Center constrains top to min and min to max
Basline constrains the top to the max value offset by the baseline

Horizontal constrains depending on the right-to-left setting. Start and end are either the min or max bound depending on rtl.
Min is constrained to the max value
Max is constrained to the min value
Center constains min to min value and max to max value

The layout description is used to create a single BookLayout which is used to layout each book.

First the content of a  book is set to the fields in the BookLayout, which also sets the bounds of the field. Then the constraints are calculated and the fields are moved according to them. I have a place where we look for overlapping fields and resolve them, but that isn't implemented yet.