Add a lot of comments with some small refactors
The ColumnLayout uses the TextLayout setContent() method to set the bounds and baseline

Change horizontalPointsToPixels to pointsToHorizontalPixels and verticalPointsToPixels to pointsToVerticalPixels

Removed dependencies from the PageLayoutHandler.

Moved layout clear() call to calculateLayout in PageLayoutHandler.

Change alignBounds to properly handle unspecified top alignment.

Changed the layout initialization in PDFPrinter to be more encapsulated