Fix up field widths and heights
When I added the "By:" field, it pushed the authors past the column end. I needed to be able to drop the field width to push it back inside the column. I was using convention where specifying both the start and end/top and bottom would center the field between the dimensions. I decided to handle centering separately and use the start, end/top, bottom specification to set the field width/height.

Since I needed to handle changing the layout width, which required creating a new layout, I changed the DynamicLayout to a StaticLayout and create each time I set the contents. I looked at the code and it looked like the DynamicLayout was doing this anyway, so I don't think it will be a big change.

I combined the horizontal and vertical layouts into a single set, and use a sealed interface to guarantee that they will be consistent.

I separated the horizontal and vertical layout calculations, to allow dependency cycles as long as they aren't both horizontal or vertical.

The orphan handler also wasn't handling fields on the same line properly, so I fixed that, too.

I moved the layout description to a separate file. This helped encapsulate some of the localization and construction.