Add thumbnail to print
I tried loading the thumbnail in setContent() but that made the layout too long, so I just keep the bookId and load the thumbnail in draw. This meant that several methods needed to be changed to suspended.

If a thumbnails doesn't fit in a column, then it needs to go to the next column. I needed to add code to handle that. I added new method to DrawLayout to return the position where a layout should be broken. This means the size needs to be specified in the field description or the bottom must be constrained by an alignment.

I needed to fix problems in handling margins. The rule is now that LayoutDimension will return values that include margins and the alignment code works with that.