Added print preview
The print preview is a recycler view of images. The images are drawn from the page layouts into bitmaps. Pretty straight forwared, except for handling multiple coroutines drawing different pages at the same time. This was a problem because the single PageLayoutHandler in the PDFPrinter shares the same layouts for all books. I removed the pageLayoutHandler from the PDFPrinter and create them as needed. Because of this, the BookPosition and SeparatorPosition classes are no longer inner classes of the PageLayoutHandler. I also pass a PageLayoutHandler to the draw methods so they can layout a book for printing.

I moved the print parameters into a view that expands and collapses at the top of the fragment.

I added coroutine cancellation checks in several places.

I had to respond to changes in the book filter to update the preview.