Fix some bugs laying out books on small page sizes
If the size is narrow enough, some fields got a negative width and crashed in the StaticLayout. I changed the code to skip fields that are narrower than the minimum field width.

The icon bitmap could get a large height, because it was tied to the bottom of the authors. I didn't allow the field to break and caused problems when the field height exceeded the page height. I just fixed the field height to 0.5 inches. This could still be an issue if the page size gets really small.