Add new fields and layout
I added all of the fields to the layout and added an additional layout for wider columns. The wider column layout breaks the fields except, title, subtitle and description into two columns.

I changed the layout description code quite a bit to condense it. I added a generator class that creates the FieldLayoutDescriptions for all of the fields. Then a lambda callback is called to setup the layout alignments. Finally the list of fields is returned.

I noticed that the label/field pair is used a lot so I made a class to hold the pair. The class can handle labels that precede or follow the data, like 10 Pages or By: Piers Anthony. The pairs also have methods to setup the label/field for the simple or two column layout.

Do get the two column layout working, I noticed that all of the alignment calculations were based on the max and min of the dimensions the layout is based on. I realized that max, min and center were all just interpolations, 1.0, 0.0 and 0.5, of the max and min values. I added an interpolation value to select how the max and min values are compbined. I also added an offset value to offset the alignment by a fixed value. You can use the interpolation value to offset by a proportion, Finally for vertical alignment, I added a baseline boolean to offset by the baseline.