Add Series to the database
The Series is in a new table. It holds the id of the series and the title of the series. It is connected to the books table using a relationship for the database id of the series row. I also added the order of a book in a series to the books table.

I also decided to remove the books_isbn column from the database. It is a lot of code to copy the tables around, but probably is the right thing to do.

Left to do: import/export, printing, filtering, and loading series for books already in the database.