Revamp Google books login
I found out that the API to get series info require authorization on google books. So I decided to login to google books when navigating to the scan destination. I added a login/logout item to the books options menu.

I also decided to changed the LaunchForResult interface to a BooksAuthorization interface which handles the OAuth authorization directly. It also handles refreshing tokens as needed. This interface is passed to the query functions in GoogleBookLookup.

The OAuth object is a static object in the main activity that is initialized when the activity is created.