Add a fragment to lookup series of books
This fragment is only used to updated old databases that were started before BiblioTech supported series. I decided to use a new flag names SERIES to mark books that have been updated. This fragment is accessed from the BooksFragment options menu if the are any books that don't have the SERIES flag set. I also added a debug build only options menu item to clear the series flag for testing.

I also changed the name of the flag field in SeriesEntity to flags for consistency.

I added a PRESERVE value to the flag field. This marks flags that are to be preserved when deleting and updating a book. Mostly 0, except for BookEntity, which preserves the SERIES flag. I added this value to the TableDescription class and add two general purpose methods to BookDatabase to set the hidden flag for an entity. The general purpose methods ensure the preserve flag values are preserved.

I added an extension method to randomly set flags in a mask. I use this to randomly set the SELECTED, EXPANDED and SERIES flags for BookEntity.

While testing I found a bug in Undo/Redo. The undo state was not properly finished when a job was canceled with undo active. I needed to run the finally clause in withUndo() in a non-cancellable context.

I removed or commented out several Log calls.