Fix tests to preserve SERIES flag
I found out that the my tests weren't testing whether the SERIES flags was preserved. I made some changes to beef that up a bit.

I added code to change the flags flags during update tests and increased the number of updates to make sure a reasonable set of flags were changed.

I changed the code that clears the flags when updating and deleting books to preserve the PRESERVE flags.

I fixed a bug in nextFlags to properly randomize over the full set of flags.