Fixes for release build
I needed to add an OAuth id for the release build because its package name and fingerprint are different from the debug build. So I add new environment variables and added a _DEBUG to the old ones.

I also discovered that I could get the environment variables from the BuildConfig and removed the uses of EnvironmentValues.

I discovered that the Google Books API library uses redirection to build the request URL and I needed to keep some classes and attributes to ProGuard to keep that working.

I update the README