Add Selected column to filter code and used it implement selected only
Add a SELECTED enum to the Column enums. This is used to filter and sort by the selection state of a book. I decided to use the predicate is and isn't to filter selected or unselected books and to only allow the single value  "true" for the value. To make that work, I added a single value string to the ColumnDescriptor that indicates that the values for the filter are not editable and the single value allowed. When sorting by the selection state, the selected books are sorted ahead of unselected books. I also added a boolean to the ColumnDescriptor to disable the header switch for the column.

I added a switch to the app toolbar to force filtering selected books without updating the filter saved for the current list. This switch completely overrides the current filter and so you really can't filter only unselectred books. When switching to a view or setting a view in the print and export fragments, the switch is set to the filtering specified in the filter itself.

With the switch, I decided to move the UI code for the switch and the book and selected counts into the MainActivity. An interface, BookStats, is used to connect the MainActivity to the filter in each fragment.

I enabled view binding and use it in the MainActivity. I will switch to view binding for other fragment eventually.

I reverted the flag I added to filter selected books.

I added a SelectionType enum that is used to select the override for selection filtering and to specify how a filter handles selected books. An extension methos, filterForSelectionType, builds a filter from another filter so the new filter filters selected books as specificed by the selection type requested.

I added an interface and class, FiltersAndCounted and FilteredBookCount, to handle overriding the selection type of the current filter. You give it a view name and a selection type, and it creates the filter and built filter for the specification. It also uses the built filter to keep count of the number of filtered and selected books. A status flow is provided that holds the filters. Methods are provided to transform this flow as needed and to get the most recent value. The interface is used to by the MainActivity to report stats and to update the selection type from the switch in the app bar.