Mostly Comments and tests and some automatic refactoring for warnings
Added comments to the code and tests for the new SELECTED column.

A few small other changes:

Force two different calls to makeBook() to use the same time to guarantee that the books are the same.

Hide the book stats when there isn't a source.

Use the "*" selector if not select values are given in buildFilter()

Moved the BookStats calls in ScanFragment to onPause and onResume to match the other fragments.