﻿Bookshelves work in progress
Export database schema so we can use the automigrate feature.

Fixed warning in BookFilterTest. runBlocking within another runBlocking scope. I changed several of the test methods to be suspend and made sure there was a runBlocking higher in the call tree.

Fixed unused warning in BookTagDaoTest.

Added Bookshelves table with migration to the database and undo/refo

Fixed some incorrect comments in UndoRedo.

Fixed a redundant ?.let in SeriesDao.