Save bookshelves work in progress
I just realized that the Google books API doesn't include creating, removing or editing bookshelves - only adding and removing volumes. So I need to rethink how Biblio-Tech will use bookshelves. But I want to keep the UI for in case I ever need to resurect it.

My thinking at this time is that I will keep a bookshelf table in the database, so I can track changes made to bookshelves on the books.google.com website. The UI will be its on fragment of Biblio-Tech and will allow syncing bookshelves with tags in Biblio-Tech. The new behavior will work like this:

First, the changes in the names and descriptions of the bookshelves will be synchronized. Removed bookshelve will be noted as removed but will not be deleted.

Then a list of the current bookshelves will be presented including the removed shelves. The user can select the bookshelves to synchronize and the synchronize them.

Existing bookshelves that have been synced before will be updated to make the tagged books in Biblio-Tech match the volumes in the bookshelf. Volumes will be added and removed from the bookshelf and be tagged and untagged. Optionally the books removed from the bookshelf are removed from Biblio-Tech, but only if the only tag left on the book is the "IoB" tag indicating the book was imported from a bookshelf..

When syncing new bookshelves, a tag will be created or linked. All volumes in the shelf will be added to Biblio-Tech and tagged with the bookshelf tag and with a tag, "IoB" indicating that the book was imported from at least one bookshelf.

When syncing a removed bookshelf, The books are treated as though all of the books were removed from the bookshelf, and then the tag for the bookshelf is deleted.

These ideas will change as I get this going.