﻿Bookshelves Work in Progress
Added get, add, update, remove methods to BookRepository

Added a new field to the book shelves table that tracks the volumes modified value the last time the shelf volumes were refreshed. This will be used to skip volume updates when the time hasn't changed. Removed some flags that couldn't be supported. Added methods to handle deleting tags that are linked to book shelves.

Changed the name of the flag that indicates a tag is linked to a bookshelf. Also called the bookshelves dao to clear tages when they are deleted.

Added code to the Google book lookup to get the current list of shelves.

Added a flag when changing the slection of tags to prevent editing tags linked to book shelves.

Got rid of editing and deleting bookshelves in the bookshelves fragment. Don't need selection either, since the shelf list just trackes the one in Google books.

Added methods to link and unlink tags and bookshelves and to refresh bookshelves from Google books.

Changed the view tag for item view in the tags recycler to hold the tag insted of the id, so I could check whether the tag is editable.

Remove the check mark icon because we aren't using it.

Adde refresh icon.