Bookshelves Work in Progress
I fetched a list of shelves from google and added them to the database, but the display isn't working.

Changed BookshelfEntity to allow null for description and selfLink. Ignore shelves with null title or id.

Setup the bookshelves fragment menu

Changed the forceBookRefresh so we still will not refresh volumes if no tag is linked.