Change modTime from and Long? to ByteArray?
I renamed it to extraData because it may have other information for other undo operations.

This will allow me to store strings for undoing updates of bookshelf books.

I call the BookDatabase_AutoMigration_8_9_Impl.migrate() method from the manual migration. You are not supposed to do this, but it works. May cause upgrade problems.